requirements.txt
reportlab>=4.0.0
pypdf>=3.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
requests>=2.31.0

